requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
chromadb>=0.5.0
openai>=1.50.0
streamlit>=1.38.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
langsmith>=0.1.0
langchain-text-splitters>=0.3.0
unstructured>=0.15.0
markdown>=3.5.0